add persistent peerStore, move related code bits to storage.go
